feat(async-map): better Promise.all + map (#3422)
