+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 
 ### Released packages
 
+- @xen-orchestra/async-map v0.0.0
 - @xen-orchestra/mixin v0.0.0
 - @xen-orchestra/emit-async v0.0.0
 - xo-server v5.27.0
