+++ packages/xo-server/package.json
@@ -32,6 +32,7 @@
   },
   "dependencies": {
     "@babel/polyfill": "7.0.0",
+    "@xen-orchestra/async-map": "^0.0.0",
     "@xen-orchestra/cron": "^1.0.3",
     "@xen-orchestra/emit-async": "^0.0.0",
     "@xen-orchestra/fs": "^0.3.0",
