+++ packages/xo-server/src/xo-mixins/jobs/index.js
@@ -2,13 +2,14 @@
 
 import type { Pattern } from 'value-matcher'
 
+import asyncMap from '@xen-orchestra/async-map'
 import { CancelToken, ignoreErrors } from 'promise-toolbox'
 import { map as mapToArray } from 'lodash'
 import { noSuchObject } from 'xo-common/api-errors'
 
 import Collection from '../../collection/redis'
 import patch from '../../patch'
-import { asyncMap, serializeError } from '../../utils'
+import { serializeError } from '../../utils'
 
 import type Logger from '../logs/loggers/abstract'
 import { type Schedule } from '../scheduling'
