+++ packages/xo-server/src/xo-mixins/scheduling.js
@@ -1,12 +1,12 @@
 // @flow
 
+import asyncMap from '@xen-orchestra/async-map'
 import { createSchedule } from '@xen-orchestra/cron'
 import { keyBy } from 'lodash'
 import { noSuchObject } from 'xo-common/api-errors'
 
 import Collection from '../collection/redis'
 import patch from '../patch'
-import { asyncMap } from '../utils'
 
 export type Schedule = {|
   cron: string,
